Upgrade clap

+++ src/main.rs
@@ -8,7 +8,7 @@ mod schema;
 
 use crate::binance::{KlineQuery, MarketEndpoint};
 use chrono::{DateTime, Utc};
-use clap::{AppSettings, Parser, Subcommand};
+use clap::{Parser, Subcommand};
 use diesel::pg::PgConnection;
 use diesel::prelude::*;
 use log::warn;
@@ -21,20 +21,18 @@ fn main() {
 
 #[derive(Parser)]
 #[clap(author, version, about)]
-#[clap(global_setting(AppSettings::PropagateVersion))]
-#[clap(global_setting(AppSettings::UseLongFormatForHelpSubcommand))]
-#[clap(setting(AppSettings::SubcommandRequiredElseHelp))]
+#[clap(propagate_version = true, subcommand_required = true, arg_required_else_help = true)]
 struct Cli {
     /// Silence all output
-    #[clap(short, long)]
+    #[clap(short, long, action)]
     quiet: bool,
 
     /// Increase message verbosity
-    #[clap(short, long, parse(from_occurrences))]
-    verbose: usize,
+    #[clap(short, long, action = clap::ArgAction::Count)]
+    verbose: u8,
 
     /// Timestamp (sec, ms, ns, none)
-    #[clap(short, long)]
+    #[clap(short, long, arg_enum, value_parser)]
     timestamp: Option<stderrlog::Timestamp>,
 
     #[clap(subcommand)]
@@ -46,7 +44,7 @@ impl Cli {
         stderrlog::new()
             .module(module_path!())
             .quiet(self.quiet)
-            .verbosity(self.verbose)
+            .verbosity(self.verbose as usize)
             .timestamp(self.timestamp.unwrap_or(stderrlog::Timestamp::Off))
             .init()
             .unwrap();
@@ -60,7 +58,7 @@ enum Commands {
     /// Snapshot data to database
     Snapshot {
         /// The database to store
-        #[clap(short, long)]
+        #[clap(short, long, value_parser)]
         database_url: String,
 
         /// Snapshot data to database
@@ -105,65 +103,65 @@ enum BinanceCommands {
     /// Fetch history Klines
     Kline {
         /// Choose a market
-        #[clap(short, long, arg_enum)]
+        #[clap(short, long, arg_enum, value_parser)]
         market: MarketEndpoint,
 
         /// The CSV file containing tasks of sync
-        #[clap(short, long)]
+        #[clap(short, long, value_parser)]
         csv: String,
 
         /// Use the interval instead of interval in CSV
-        #[clap(short, long)]
+        #[clap(short, long, value_parser)]
         interval: Option<String>,
 
         /// Use the limit instead of limits in CSV
-        #[clap(short, long)]
+        #[clap(short, long, value_parser)]
         limit: Option<u16>,
 
         /// Start time
-        #[clap(long = "from")]
+        #[clap(long = "from", value_parser)]
         start_time: Option<DateTime<Utc>>,
 
         /// End time
-        #[clap(long = "to")]
+        #[clap(long = "to", value_parser)]
         end_time: Option<DateTime<Utc>>,
     },
 
     /// Watch Klines in real time
     KlineStream {
         /// Choose a market
-        #[clap(short, long, arg_enum)]
+        #[clap(short, long, arg_enum, value_parser)]
         market: MarketEndpoint,
 
         /// The CSV file containing tasks of sync
-        #[clap(short, long)]
+        #[clap(short, long, value_parser)]
         csv: String,
 
         /// Use the interval instead of interval in CSV
-        #[clap(short, long)]
+        #[clap(short, long, value_parser)]
         interval: Option<String>,
     },
 
     /// Fetch open interest summaries
     OpenInterestSummary {
         /// The CSV file containing tasks of sync
-        #[clap(short, long)]
+        #[clap(short, long, value_parser)]
         csv: String,
 
         /// Use the interval instead of interval in CSV
-        #[clap(short, long)]
+        #[clap(short, long, value_parser)]
         interval: Option<String>,
 
         /// Use the limit instead of limits in CSV
-        #[clap(short, long)]
+        #[clap(short, long, value_parser)]
         limit: Option<u16>,
 
         /// Start time
-        #[clap(long = "from")]
+        #[clap(long = "from", value_parser)]
         start_time: Option<DateTime<Utc>>,
 
         /// End time
-        #[clap(long = "to")]
+        #[clap(long = "to", value_parser)]
         end_time: Option<DateTime<Utc>>,
     },
 }
